// src/api/generated/models/reviewAuditRequestStatus.ts
/**
 * Generated by orval v8.7.0 🍺
 * Do not edit manually.
 * OpenAPI definition
 * OpenAPI spec version: v0
 */

export type ReviewAuditRequestStatus = typeof ReviewAuditRequestStatus[keyof typeof ReviewAuditRequestStatus];


export const ReviewAuditRequestStatus = {
  DRAFT: 'DRAFT',
  PENDING: 'PENDING',
  APPROVED: 'APPROVED',
  REJECTED: 'REJECTED',
  WITHDRAWN: 'WITHDRAWN',
} as const;